Move IndieAuth-related handler stuff to own package
